Extracted write methods into the new Write module.
