+++ lib/honey.ex
@@ -1,7 +1,6 @@
 defmodule Honey do
-  alias Honey.Directories
-  alias Honey.Boilerplates
   alias Honey.{Translator, Utils, Fuel}
+  alias Honey.Write
 
   @moduledoc """
   Honey Potion is a framework that brings the powerful eBPF technology into Elixir.
@@ -48,55 +47,6 @@ defmodule Honey do
   end
 
   @doc """
-  Writes c_code into a file in proj_path with module_name.
-  This can use an optional clang_format with the last parameter.
-  """
-
-  def write_c_file(c_code, proj_path, module_name, clang_format) do
-    "Elixir." <> module_name = "#{module_name}"
-
-    c_path =
-      proj_path
-      |> Path.dirname()
-      |> Path.join("src/#{module_name}.bpf.c")
-
-    {:ok, file} = File.open(c_path, [:write])
-    IO.binwrite(file, c_code)
-    File.close(file)
-
-    # Optional, format the file using clang-format:
-    clang_format && System.cmd(clang_format, ["-i", c_path])
-
-    true
-  end
-
-  @doc """
-  Prepares a generic makefile in the directory of the user, writes a generic front-end for eBPF and compiles everything.
-  This uses the LibBPF located in /benchmarks/libs/
-  """
-
-  def compile_eBPF(proj_path, module_name) do
-
-    userdir = proj_path |> Path.dirname()
-
-    makedir = Path.join(:code.priv_dir(:honey), "BPF_Boilerplates/Makefile")
-    File.cp_r(makedir, userdir |> Path.join("Makefile"))
-
-    mod_name = Atom.to_string(module_name)
-    mod_name = String.slice(mod_name, 7, String.length(mod_name) - 7)
-
-    frontend = Boilerplates.generate_frontend_code(mod_name)
-    File.write(userdir |> Path.join("./src/#{mod_name}.c"), frontend)
-
-    libsdir = __ENV__.file |> Path.dirname
-    libsdir = Path.absname("./../benchmarks/libs/libbpf/src", libsdir) |> Path.expand
-
-    privdir = :code.priv_dir(:honey)
-
-    System.cmd("make", ["TARGET := #{mod_name}", "LIBBPF_DIR := #{libsdir}", "PRIV_DIR := #{privdir}"], cd: userdir)
-  end
-
-  @doc """
   Macro that allows users to define maps in eBPF through elixir.
   Users can define maps using the macro defmap. For example, to create a map named my_map, you can:
 
@@ -119,10 +69,11 @@ defmodule Honey do
   end
 
   @doc """
-  Honey-Potion runs using the __before_compile__ macro. This macro starts up the execution.
+  Honey-Potion runs using the __before_compile__ macro. So here is where we keep the Honey-Potion pipeline. 
   """
 
   defmacro __before_compile__(env) do
+    #This can go into a new "Guard" module.
     target_func = :main
     target_arity = 1
     #If the main function isn't defined raise an error.
@@ -133,9 +84,12 @@ defmodule Honey do
       )
     end
 
+    #This can go into Utils as ExtractAst or into Info with same name.
     # TODO: evaluate all clauses
     {:v1, _kind, _metadata, [first_clause | _other_clauses]} = main_def
     {_metadata, arguments, _guards, func_ast} = first_clause
+
+    #This can become a one-line call.
     #Burns Fuel (expands recursive calls into repetitions) and runs the optimizer on the AST.
     final_ast =
       func_ast
@@ -144,6 +98,7 @@ defmodule Honey do
 
     # print_ast(final_ast)
 
+    #This can go into Utils as ExtractEnv or into Info.
     ebpf_options = Module.get_attribute(env.module, :ebpf_options)
     #Gets values required to translate the AST to eBPF readable C.
     sections = Module.get_attribute(env.module, :sections)
@@ -152,20 +107,15 @@ defmodule Honey do
     maps = Module.get_attribute(env.module, :ebpf_maps)
     # TODO: env.requires stores the requires in alphabetical order. This might be a problem.
 
+    clang_format = Keyword.get(ebpf_options, :clang_format)
+
     #Calls the code translator.
     c_code = Translator.translate("main", final_ast, sec, license, env.requires, maps)
 
-    #Creates the directories for the file and compilation.
-    Directories.create_all(env.file)
-
-    #Writes the file.
-    clang_format = Keyword.get(ebpf_options, :clang_format)
-    write_c_file(c_code, env.file, env.module, clang_format)
+    #Writes .bpf.c and .c files for each program in the lib folder.
+    Write.write_ouput_files(c_code, env.file, env.module, clang_format)
 
     Module.delete_definition(env.module, {target_func, target_arity})
-
-    #Does the compilation to an eBPF executable
-    compile_eBPF(env.file, env.module)
 
     quote do
       def main(unquote(arguments)) do
